Rename boss-dtos to boss-transfers
