+++ boss-services/src/main/java/org/ssor/boss/service/UserService.java
@@ -6,11 +6,11 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
-import org.ssor.boss.dto.ApiRequestResponse;
-import org.ssor.boss.dto.RegisterUserInput;
-import org.ssor.boss.dto.RegisterUserOutput;
-import org.ssor.boss.dto.SecureUserDetails;
-import org.ssor.boss.dto.UpdateUserInput;
+import org.ssor.boss.transfer.ApiRequestResponse;
+import org.ssor.boss.transfer.RegisterUserInput;
+import org.ssor.boss.transfer.RegisterUserOutput;
+import org.ssor.boss.transfer.SecureUserDetails;
+import org.ssor.boss.transfer.UpdateUserInput;
 import org.ssor.boss.entity.User;
 import org.ssor.boss.repository.UserRepository;
 import javax.validation.Valid;
@@ -78,7 +78,7 @@ public class UserService implements UserDetailsService
    */
   public ApiRequestResponse updateUserProfile(@Valid @NotNull UpdateUserInput updateUserInput)
   {
-
+    return null;
   }
 
   /**
