+++ boss-services/src/test/java/org/ssor/boss/service/UserServiceTests.java
@@ -9,9 +9,9 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.ssor.boss.dto.RegisterUserInput;
-import org.ssor.boss.dto.RegisterUserOutput;
-import org.ssor.boss.dto.SecureUserDetails;
+import org.ssor.boss.transfer.RegisterUserInput;
+import org.ssor.boss.transfer.RegisterUserOutput;
+import org.ssor.boss.transfer.SecureUserDetails;
 import org.ssor.boss.entity.User;
 import org.ssor.boss.repository.UserRepository;
 import java.time.LocalDateTime;
