+++ boss-transfers/pom.xml
@@ -8,7 +8,7 @@
     <version>0.0.1-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-  <artifactId>boss-dtos</artifactId>
+  <artifactId>boss-transfers</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <properties>
     <spring-boot.repackage.skip>true</spring-boot.repackage.skip>
