+++ boss-transfers/src/main/java/org/ssor/boss/transfer/SecureUserDetails.java
@@ -1,4 +1,4 @@
-package org.ssor.boss.dto;
+package org.ssor.boss.transfer;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
