+++ boss-transfers/src/test/java/org/ssor/boss/transfer/RegisterUserOutputTests.java
@@ -1,4 +1,4 @@
-package org.ssor.boss.dto;
+package org.ssor.boss.transfer;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
